--- a/test/Microsoft.AspNetCore.Server.Kestrel.Performance/project.json
+++ b/test/Microsoft.AspNetCore.Server.Kestrel.Performance/project.json
@@ -8,7 +8,7 @@
     "netcoreapp1.0": {
       "dependencies": {
         "Microsoft.NETCore.App": {
-          "version": "1.0.1-*",
+          "version": "1.1.0-*",
           "type": "platform"
         }
       }