Results on 16 core (32 HT) Server

> BenchmarkDotNet=v0.10.0
OS=Windows
Processor=?, ProcessorCount=32
Frequency=2343747 Hz, Resolution=426.6672 ns, Timer=TSC
Host Runtime=.NET Core 4.6.24628.01, Arch=64-bit  [RyuJIT]
GC=Concurrent Server
dotnet cli version=1.0.0-preview2-1-003177
Job Runtime(s):
        .NET Core 4.6.24410.01, Arch=64-bit  [RyuJIT]

> RemoveOutliers=False  Runtime=Core  Server=True
LaunchCount=3  RunStrategy=Throughput  TargetCount=10
WarmupCount=5

Single threaded
```
                   Method |          Mean |      StdDev | Scaled |           RPS |
------------------------- |-------------- |------------ |------- |-------------- |
           ParsePlaintext |   993.3004 ns |  23.2932 ns |   1.00 |  1,006,744.82 |
  ParsePipelinedPlaintext |   752.0448 ns |  15.0722 ns |   0.76 |  1,329,708.02 |
          ParseLiveAspNet | 4,347.4022 ns |  81.1657 ns |   4.38 |    230,022.43 |
 ParsePipelinedLiveAspNet | 4,026.5624 ns |  60.3856 ns |   4.06 |    248,350.80 |
             ParseUnicode | 6,859.2000 ns | 147.5681 ns |   6.91 |    145,789.60 |
    ParseUnicodePipelined | 6,679.8011 ns | 210.6329 ns |   6.73 |    149,705.06 |
```
Multi threaded
```
                   Method |          Mean |      StdDev | Scaled |           RPS |
------------------------- |-------------- |------------ |------- |-------------- |
           ParsePlaintext |   188.0574 ns |   1.5506 ns |   1.00 |  5,317,526.25 |
  ParsePipelinedPlaintext |    67.4741 ns |   1.9620 ns |   0.36 | 14,820,495.27 |
          ParseLiveAspNet |   576.7441 ns |   3.2424 ns |   3.07 |  1,733,871.11 |
 ParsePipelinedLiveAspNet |   309.2693 ns |  11.8032 ns |   1.64 |  3,233,428.06 |
             ParseUnicode |   727.9797 ns |  23.8009 ns |   3.87 |  1,373,664.74 |
    ParseUnicodePipelined |   493.8310 ns |  21.8476 ns |   2.63 |  2,024,984.33 |
```