@benaadams 2016-11-22
@stephentoub is this the correct way to use Parallel.For for maximum throughput? File `MultiThreadedRequestParsing.cs`