@stephentoub 2016-11-22
> is this the correct way

I expect that'll be about as good as you can get with it, assuming nothing else is taking up thread pool threads while this runs.  There will be a small amount of overhead in getting started, but it shouldn't be measurable assuming enough work is done in the body.