@benaadams 2016-11-22
4 Core desktop

> BenchmarkDotNet=v0.10.0
OS=Windows
Processor=?, ProcessorCount=4
Frequency=2825619 Hz, Resolution=353.9048 ns, Timer=TSC
Host Runtime=.NET Core 4.6.24628.01, Arch=64-bit  [RyuJIT]
GC=Concurrent Server
dotnet cli version=1.0.0-preview2-1-003155
Job Runtime(s):
        .NET Core 4.6.24410.01, Arch=64-bit  [RyuJIT]

> RemoveOutliers=False  Runtime=Core  Server=True
LaunchCount=3  RunStrategy=Throughput  TargetCount=10
WarmupCount=5

Single threaded
```
                   Method |          Mean |     StdDev | Scaled |          RPS |
------------------------- |-------------- |----------- |------- |------------- |
           ParsePlaintext |   793.1210 ns | 13.8271 ns |   1.00 | 1,260,841.62 |
  ParsePipelinedPlaintext |   626.8362 ns |  6.6949 ns |   0.79 | 1,595,313.01 |
          ParseLiveAspNet | 3,842.9023 ns | 46.0246 ns |   4.85 |   260,219.99 |
 ParsePipelinedLiveAspNet | 3,542.9391 ns | 44.6045 ns |   4.47 |   282,251.53 |
             ParseUnicode | 6,459.0080 ns | 33.4147 ns |   8.15 |   154,822.54 |
    ParseUnicodePipelined | 6,237.0520 ns | 67.3694 ns |   7.87 |   160,332.16 |
```

Multi threaded
```
                   Method |          Mean |     StdDev | Scaled |          RPS |
------------------------- |-------------- |----------- |------- |------------- |
           ParsePlaintext |   295.8569 ns |  3.4383 ns |   1.00 | 3,380,012.21 |
  ParsePipelinedPlaintext |   184.1479 ns |  4.0847 ns |   0.62 | 5,430,417.46 |
          ParseLiveAspNet | 1,238.1511 ns | 41.2725 ns |   4.19 |   807,655.84 |
 ParsePipelinedLiveAspNet | 1,039.3710 ns | 17.6234 ns |   3.51 |   962,120.36 |
             ParseUnicode | 1,985.8980 ns | 33.2729 ns |   6.71 |   503,550.54 |
    ParseUnicodePipelined | 1,825.3181 ns | 30.2562 ns |   6.17 |   547,849.71 |
```