@benaadams 2016-12-04
Updated for BenchmarkDotNet to 0.10.1

@halter73 @cesarbs @pakrym worth the upgrade; is 1.1.0 and new memory diag built in

New output (minus a few columns)
```
                   Method |          Mean | Scaled |          RPS | Allocated |
------------------------- |-------------- |------- |------------- |---------- |
           ParsePlaintext |   849.5240 ns |   1.00 | 1,177,129.73 |     116 B |
  ParsePipelinedPlaintext |   647.2198 ns |   0.76 | 1,545,070.11 |     104 B |
          ParseLiveAspNet | 3,918.9647 ns |   4.61 |   255,169.44 |   1.11 kB |
 ParsePipelinedLiveAspNet | 3,600.5674 ns |   4.24 |   277,734.00 |    1.1 kB |
             ParseUnicode | 6,751.1812 ns |   7.95 |   148,122.23 |   1.94 kB |
    ParseUnicodePipelined | 6,394.2506 ns |   7.53 |   156,390.49 |   1.93 kB |
```