@pakrym 2016-12-06
@benaadams which part of kestrel code does this benchmark vs single-threaded version?