@benaadams 2016-12-06
@pakrym same code, just maxes it out across cores rather than single core; it should be embarrassingly parallel; though the increase I am seeing isn't as much as x N cores - which may be interesting