@halter73 2016-12-06
This doesn't test any kind of contention. It might be more useful to have a test that keeps the producer going during consumption to test contention acquiring SocketInput's _sync lock.

Maybe you could use the threadpool to dispatch calls to InsertData and ensure that there is enough data always in the buffer to prevent the consumer from ever catching up with the producer. By looking at the return values of TakeStartLine and TakeMessageHeaders it should be possible to see if the consumer ever stalls.